refactor Supplement file structure
- pulled classes out of files with multiple classes to conform to C# standards and to help with location and readability.
- Supplement.cs also now overrides OnCreated() to hide fodder storetypes which historically are not mutatable by end users.
